requirements.txt
flask
flask-cors

# LangChain + LangGraph ecosystem
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
faiss-cpu

python-dotenv
